FINERACT-1724: Compilation error fix

+++ integration-tests/src/test/java/org/apache/fineract/integrationtests/ExternalIdSupportIntegrationTest.java
@@ -934,7 +934,7 @@ public class ExternalIdSupportIntegrationTest extends IntegrationTest {
             PutLoansLoanIdResponse modifyLoanApplicationResult = this.loanTransactionHelper.modifyLoanApplication(loanExternalIdStr2,
                     "modify",
                     new PutLoansLoanIdRequest().submittedOnDate("31 August 2022").dateFormat("dd MMMM yyyy").locale("en")
-                            .loanType("individual").productId(loanProductID).clientId(client.getClientId()).interestType(0)
+                            .loanType("individual").productId(loanProductID.longValue()).clientId(client.getClientId()).interestType(0)
                             .interestCalculationPeriodType(1).interestRatePerPeriod(0).isEqualAmortization(false).loanTermFrequency(30)
                             .loanTermFrequencyType(0).maxOutstandingLoanBalance(10000L).numberOfRepayments(1).principal(10000L)
                             .repaymentEvery(30).repaymentFrequencyType(0).transactionProcessingStrategyCode("mifos-standard-strategy")
